implement simple allocator

+++ 03.allocator/allocator.cpp
@@ -9,7 +9,7 @@
 
 
 int factorial(const int n) {
-    long f = 1;
+    int f = 1;
     for (int i = 1; i <= n; ++i)
         f *= i;
     return f;
